Improved "select" type.
Removed "select-search" type.
"select" now support groups with headers.
"search" and "fullTextSearch" attributes (booleans) can now be used to make a "select" searchable.
"placeholder" attribute (string) can be used to change the default placeholder of a "select".
An optional "select" can be cleared using the button at the top of the list.
Updated README.

Thanks to @abernix for the help. Closes #3