Fixed the height of empty date/datetime fields on iOS.
This is a tough one becasue Semantic doesn't really have official
date/datetime support, but without this on iOS the heights of these
fields when they are empty are completely collapsed.  Once they are
filled with a value, they are fine.  This just sets a min-height so
they are consistent with the other elements until the date is selected.