Placeholder should return when clearing an optional field that had an existing value.
When Clearing an Optional Field, if the form was an "insert" it was properly cleared,
however, if it was an "update" and had a value already, the placeholder was not
properly returned to it's placeholder state.  The input value WAS being updated, but the
display text for the select was left as the previous value.  This seems to fix it.