Allow the "class" to be passed through to the UI Dropdown
This allows the passing of things like `compact`, `fluid`, or `floated` through
to the UI Dropdown containing DIV element by simply passing `class=blah` on the
`afQuickField`.  Normally the `class` would be passed through to the `input`
(which is `type="hidden"` and doesn't need `class`), but this provides
the option to use the more complete scope of `ui dropdown` options.

Also, since the removal of the hardcoded `fluid` class in 0.6.1, this is
sometimes necessary so you can do `class="fluid"`.